refactor: update imports to use shadcn directory in issues pages
